refactor: cleanup

+++ app/src/engine/process/process_win.hpp
@@ -83,16 +83,16 @@ class Process : public IProcess {
 
     [[nodiscard]] bool alive() const noexcept override {
         assert(is_initalized_);
+
         DWORD exitCode = 0;
         GetExitCodeProcess(pi_.hProcess, &exitCode);
+
         return exitCode == STILL_ACTIVE;
     }
 
     void setAffinity(const std::vector<int> &cpus) noexcept override {
         assert(is_initalized_);
-        if (!cpus.empty()) {
-            affinity::setAffinity(cpus, pi_.hProcess);
-        }
+        if (!cpus.empty()) affinity::setAffinity(cpus, pi_.hProcess);
     }
 
     void killProcess() {
@@ -100,16 +100,11 @@ class Process : public IProcess {
 
         process_list.remove_if([this](const ProcessInformation &pi) { return pi.identifier == pi_.hProcess; });
 
-        try {
-            DWORD exitCode = 0;
-            GetExitCodeProcess(pi_.hProcess, &exitCode);
+        DWORD exitCode = 0;
+        GetExitCodeProcess(pi_.hProcess, &exitCode);
 
-            if (exitCode == STILL_ACTIVE) {
-                UINT uExitCode = 0;
-                TerminateProcess(pi_.hProcess, uExitCode);
-            }
-
-        } catch (const std::exception &e) {
+        if (exitCode == STILL_ACTIVE) {
+            TerminateProcess(pi_.hProcess, 0);
         }
 
         is_initalized_ = false;
@@ -133,67 +128,56 @@ class Process : public IProcess {
 
         auto id = std::this_thread::get_id();
 
-        auto readFuture = std::async(std::launch::async, [this, &last_word, &lines, id, &atomic_stop = atomic::stop]() {
-            char buffer[4096];
-            DWORD bytesRead;
+        auto read_future =
+            std::async(std::launch::async, [this, &last_word, &lines, id, &atomic_stop = atomic::stop]() {
+                char buffer[4096];
+                DWORD bytes_read;
 
-            while (true) {
-                if (!ReadFile(child_std_out_, buffer, sizeof(buffer), &bytesRead, nullptr)) {
-                    return Status::ERR;
-                }
-
-                if (atomic_stop) {
-                    return Status::ERR;
-                }
-
-                if (bytesRead <= 0) continue;
-
-                // Iterate over each character in the buffer
-                for (DWORD i = 0; i < bytesRead; i++) {
-                    // If we encounter a newline, add the current line to the vector and reset
-                    // the current_line_ on Windows newlines are \r\n. Otherwise, append the
-                    // character to the current line
-                    if (buffer[i] != '\n' && buffer[i] != '\r') {
-                        current_line_ += buffer[i];
-                        continue;
+                while (true) {
+                    if (!ReadFile(child_std_out_, buffer, sizeof(buffer), &bytes_read, nullptr)) {
+                        return Status::ERR;
                     }
 
-                    // don't add empty lines
-                    if (current_line_.empty()) continue;
+                    if (atomic_stop) return Status::ERR;
+                    if (bytes_read <= 0) continue;
 
-                    const auto time = Logger::should_log_ ? util::time::datetime_precise() : "";
+                    // Iterate over each character in the buffer
+                    for (DWORD i = 0; i < bytes_read; i++) {
+                        // If we encounter a newline, add the current line to the vector and reset
+                        // the current_line_ on Windows newlines are \r\n. Otherwise, append the
+                        // character to the current line
+                        if (buffer[i] != '\n' && buffer[i] != '\r') {
+                            current_line_ += buffer[i];
+                            continue;
+                        }
 
-                    lines.emplace_back(Line{current_line_, time});
+                        // don't add empty lines
+                        if (current_line_.empty()) continue;
 
-                    if (realtime_logging_) {
-                        Logger::readFromEngine(current_line_, time, log_name_, false, id);
+                        const auto time = Logger::should_log_ ? util::time::datetime_precise() : "";
+
+                        lines.emplace_back(Line{current_line_, time});
+
+                        if (realtime_logging_) Logger::readFromEngine(current_line_, time, log_name_, false, id);
+
+                        if (current_line_.rfind(last_word, 0) == 0) return Status::OK;
+
+                        current_line_.clear();
                     }
-
-                    if (current_line_.rfind(last_word, 0) == 0) {
-                        return Status::OK;
-                    }
-
-                    current_line_.clear();
                 }
-            }
-        });
+            });
 
         // Check if the asynchronous operation is completed
-        const auto fut = readFuture.wait_for(threshold);
+        const auto fut = read_future.wait_for(threshold);
 
-        if (fut == std::future_status::timeout) {
-            return Status::TIMEOUT;
-        }
-
+        if (fut == std::future_status::timeout) return Status::TIMEOUT;
         return Status::OK;
     }
 
     bool writeProcess(const std::string &input) noexcept override {
         assert(is_initalized_);
 
-        if (!alive()) {
-            killProcess();
-        }
+        if (!alive()) killProcess();
 
         DWORD bytesWritten;
         auto res = WriteFile(child_std_in_, input.c_str(), input.length(), &bytesWritten, nullptr);
@@ -206,6 +190,7 @@ class Process : public IProcess {
    private:
     void closeHandles() const {
         assert(is_initalized_);
+
         try {
             CloseHandle(pi_.hThread);
             CloseHandle(pi_.hProcess);
@@ -213,7 +198,7 @@ class Process : public IProcess {
             CloseHandle(child_std_out_);
             CloseHandle(child_std_in_);
         } catch (const std::exception &e) {
-            std::cerr << e.what();
+            Logger::fatal("Error closing handles: {}", e.what());
         }
     }
